first step of signing in anonymously is done
