+++ lib/styles/colors.dart
@@ -5,7 +5,7 @@ class MateColors {
   static const secondary = Color(0xffff8151);
   static const third = Color(0xff03c8bb);
   static const white = Color(0xffffffff);
-  static const grey = Color(0xff222222);
+  static const grey = Color(0xff323232);
   static const lightGrey = Color(0xaa525252);
   static const black = Color(0xff000000);
