+++ lib/views/calendar/calendar.dart
@@ -10,7 +10,7 @@ class Calendar extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final UserModel user = Provider.of<UserModel>(context);
-    if (user != null) {
+    if (!AuthService().getUser.isAnonymous) {
       return StreamBuilder(
         stream: Collection<Event>(
                 path: 'hochschulen/${user.university}/events',
