+++ lib/views/settings/local_widgets/local_widgets.dart
@@ -1,2 +1,3 @@
-export 'settings_list.dart';
+export 'settings_list_general.dart';
+export 'settings_list_information.dart';
 export 'settings_list_item.dart';
